Move thread IPC stuff into a separate module
