+++ src/common/input/handler.rs
@@ -4,7 +4,7 @@ use super::actions::Action;
 use super::keybinds::Keybinds;
 use crate::common::input::config::Config;
 use crate::common::pty::PtyInstruction;
-use crate::common::{AppInstruction, ThreadSenders, OPENCALLS};
+use crate::common::{thread_bus::OPENCALLS, AppInstruction, ThreadSenders};
 use crate::errors::ContextType;
 use crate::os_input_output::OsApi;
 use crate::screen::ScreenInstruction;
